fix: storybook base actions
